+++ Libraries/SideScroll/Logs/LogEntry.cs
@@ -53,6 +53,7 @@ public class LogSettings
 			MaxLogItems = MaxLogItems,
 			MinLogLevel = MinLogLevel,
 			DebugPrintLogLevel = DebugPrintLogLevel,
+			Context = Context, // Entries have to keep posting to the same context to stay thread safe
 		};
 	}
 
@@ -234,7 +235,14 @@ public class LogEntry : INotifyPropertyChanged
 	/// </summary>
 	protected void NotifyPropertyChanged([CallerMemberName] string propertyName = "")
 	{
-		Settings?.Context?.Post(NotifyPropertyChangedContext, propertyName);
+		if (Settings?.Context is { } context)
+		{
+			context.Post(NotifyPropertyChangedContext, propertyName);
+		}
+		else
+		{
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+		}
 	}
 
 	private void NotifyPropertyChangedContext(object? state)